requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.32.0
google-generativeai>=0.4.0
python-dotenv>=1.0.1
pandas>=2.2.0
openpyxl>=3.1.0
